% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{orderly_artefact}
\alias{orderly_artefact}
\title{Declare orderly artefacts}
\usage{
orderly_artefact(description = NULL, files)
}
\arguments{
\item{description}{The name of the artefact}

\item{files}{The files within this artefact}
}
\value{
Undefined
}
\description{
Declare an artefact. By doing this you turn on a number of orderly
features; see Details below. You can have multiple calls to this
function within your orderly script.
}
\details{
(1) files matching this will \emph{not} be copied over from the src
directory to the draft directory unless they are also listed as a
resource with \code{\link[=orderly_resource]{orderly_resource()}}. This feature is only enabled
if you call this function from the top level of the orderly script
and if it contains only string literals (no variables).

(2) if your script fails to produce these files, then
\code{\link[=orderly_run]{orderly_run()}} will fail, guaranteeing that your task does really
produce the things you need it to.

(3) within the final metadata, your artefacts will have additional
metadata; the description that you provide and a grouping
}
\examples{
# An example in context within the orderly examples:
orderly_example_show("strict")
}
