% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{orderly_build_script_details}
\alias{orderly_build_script_details}
\title{Build the details of an orderly source script}
\usage{
orderly_build_script_details(exprs, entrypoint_filename)
}
\arguments{
\item{exprs}{Parsed AST from orderly script}

\item{entrypoint_filename}{Name of entrypoint file to include in metadata}
}
\value{
Details of orderly script
}
\description{
Takes the parsed AST from an orderly script, parses details
of any calls to orderly_ in-script functions into intermediate
representation for downstream use. Also validates calls to
orderly_ in-script functions are well-formed.
}
