% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{orderly_cleanup}
\alias{orderly_cleanup}
\alias{orderly_cleanup_status}
\title{Clean up source directory}
\usage{
orderly_cleanup(name = NULL, root = NULL, locate = TRUE)

orderly_cleanup_status(name = NULL, root = NULL, locate = TRUE)
}
\arguments{
\item{name}{Name of the report directory to clean (i.e., we look
at \verb{src/<name>} relative to your orderly root}

\item{root}{The path to an orderly root directory, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}.}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{.outpack} directory}
}
\value{
An (currently unstable) object of class
\code{orderly_cleanup_status} within which the element \code{delete}
indicates files that would be deleted (for
\code{orderly_cleanup_status}) or that were deleted (for
\code{orderly_cleanup})
}
\description{
Find, and delete, file that were generated by running a report.
Until you're comfortable with what this will do, you are strongly
recommended to run \code{orderly_cleanup_status} first to see what will
be deleted.
}
\details{
After file deletion, we look through and remove all empty
directories; orderly2 has similar semantics here to git where
directories are never directly tracked.
}
\section{Notes for user of orderly1}{


In orderly1 this function has quite different semantics, because
the full set of possible files is always knowable from the yaml
file. So there, we start from the point of view of the list of
files then compare that with the directory.
}

\author{
export
}
