% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{orderly_compare_packets}
\alias{orderly_compare_packets}
\title{Compare two packets}
\usage{
orderly_compare_packets(
  target,
  current,
  search_options = NULL,
  root = NULL,
  locate = TRUE,
  what = c("metadata", "files")
)
}
\arguments{
\item{target}{The id of the packet to use in the comparison.}

\item{current}{The id of the other packet against which to compare.}

\item{search_options}{Options for locating packet files. If there are no
copies of the files locally, they can be downloaded automatically from a
remote location on-demand if \code{allow_remote} is \code{TRUE}.}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function does not require that the
directory is configured for orderly, and can be any \code{outpack}
root (see \link{orderly_init} for details).}

\item{locate}{Logical, indicating if the root should be searched
for.  If \code{TRUE}, then we looks in the directory given for \code{root}
(or the working directory if \code{NULL}) and then up through its
parents until it finds an \code{.outpack} directory or
\code{orderly_config.yml}}

\item{what}{One or more of "metadata", "files" and "artefacts", retricting
what components of the packet to compare. This is useful when it is known
for example that the source code of a report what changed, and one is only
interested in the effect on its output.}
}
\value{
An R6 object of class \code{orderly_packet_diff} is returned. Printing
this object will show the difference in the two packets. Additionally, the
\code{is_equal} method of the object return a logical indicating whether any
difference was found.
}
\description{
Compare the metadata and contents of two packets.
}
\details{
Insignificant differences in the metadata (eg. different dates and packet
IDs) are excluded from the comparison.

If either packet is not unpacked, but \code{search_options}'s \code{allow_remote} is
\code{TRUE}, we will try to request files from remote locations, as necessary.
}
