% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{orderly_dependency}
\alias{orderly_dependency}
\title{Declare a dependency}
\usage{
orderly_dependency(name, query, use)
}
\arguments{
\item{name}{The name of the packet to depend on}

\item{query}{The query to search for; often this will simply be
the string \code{latest}, indicating the most recent version. You may
want a more complex query here though.}

\item{use}{Files to use from the packet found by \code{query}, usually
as a named character vector with string interpolation in the
names; see \link{outpack_packet_use_dependency}'s \code{file}
argument for details.}
}
\value{
Undefined
}
\description{
Declare a dependency on another packet
}
\details{
See \link{orderly_run} for some details about how search
options are used to select which locations packets are found from,
and if any data is fetched over the network. If you are running
interactively, this will obviously not work, so you should use
\code{\link[=orderly_interactive_set_search_options]{orderly_interactive_set_search_options()}} to set the
options that this function will respond to.
}
