% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{orderly_depends}
\alias{orderly_depends}
\title{Declare a dependency}
\usage{
orderly_depends(name, query, use)
}
\arguments{
\item{name}{The name of the packet to depend on}

\item{query}{The query to search for; often this will simply be
the string \code{latest}, indicating the most recent version. You may
want a more complex query here though.}

\item{use}{A named character vector of filenames to copy from the
upstream packet. The name corresponds to the destination name,
so c(here.csv = "there.csv") will take the upstream file
\code{there.csv} and copy it over as \code{here.csv}.}
}
\value{
Undefined
}
\description{
Declare a dependency on another packet
}
