% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{orderly_example}
\alias{orderly_example}
\title{Copy a simple orderly example}
\usage{
orderly_example(name, ..., dest = NULL)
}
\arguments{
\item{name}{The name of the example to copy. Currently only
"default" is supported.}

\item{...}{Arguments passed through to \code{\link[=orderly_init]{orderly_init()}}}

\item{dest}{The destination. By default we use
\code{withr::local_tempfile()} which will create a temporary
directory that will clean itself up. This is suitable for use
from the orderly examples, but you may prefer to provide your
own path. The path must not already exist.}
}
\value{
Invisibly, the path to the example.
}
\description{
Copy a simple orderly example for use in the docs. This function
should not form part of your workflow!
}
\examples{
path <- orderly2::orderly_example("default")
orderly2::orderly_list_src(root = path)

fs::dir_delete(path)
}
