% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{orderly_import_zip}
\alias{orderly_import_zip}
\title{Import packets from a zip file.}
\usage{
orderly_import_zip(path, root = NULL, locate = TRUE)
}
\arguments{
\item{path}{the path to the zip file to be imported.}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function does not require that the
directory is configured for orderly, and can be any \code{outpack}
root (see \link{orderly_init} for details).}

\item{locate}{Logical, indicating if the root should be searched
for.  If \code{TRUE}, then we looks in the directory given for \code{root}
(or the working directory if \code{NULL}) and then up through its
parents until it finds an \code{.outpack} directory or
\code{orderly_config.yml}}
}
\value{
Invisibly, the IDs of the imported packets
}
\description{
Import packets from a zip file.
}
