% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{orderly_list_src}
\alias{orderly_list_src}
\title{List source reports}
\usage{
orderly_list_src(root = NULL, locate = TRUE)
}
\arguments{
\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function \strong{does} require that the
directory is configured for orderly, and not just outpack (see
\link{orderly_init} for details).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{.outpack} directory}
}
\value{
A character vector of names of source reports, suitable
for passing to \link{orderly_run}
}
\description{
List source reports - that is, directories within \verb{src/} that
look suitable for running with orderly; these will be directories
that contain an entrypoint file - a \code{.R} file with the same name
as the directory (e.g., \code{src/data/data.R} corresponds to \code{data}).
}
\examples{
path <- orderly2::orderly_example("default")
orderly2::orderly_list_src(root = path)
}
\seealso{
\link{orderly_metadata_extract} for listing packets
that have completed
}
