% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{orderly_location_list}
\alias{orderly_location_list}
\title{List known pack locations}
\usage{
orderly_location_list(verbose = FALSE, root = NULL, locate = TRUE)
}
\arguments{
\item{verbose}{Logical, indicating if we should return a
data.frame that includes more information about the location.}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function does not require that the
directory is configured for orderly, and can be any \code{outpack}
root (see \link{orderly_init} for details).}

\item{locate}{Logical, indicating if the root should be searched
for.  If \code{TRUE}, then we looks in the directory given for \code{root}
(or the working directory if \code{NULL}) and then up through its
parents until it finds an \code{.outpack} directory or
\code{orderly_config.yml}}
}
\value{
Depending on the value of \code{verbose}:
\itemize{
\item \code{verbose = FALSE}: A character vector of location names. The
special name \code{local} will always be present.  This is the
default behaviour.
\item \code{verbose = TRUE}: A data.frame with columns \code{name}, \code{type} and
\code{args}.  The \code{args} column is a list column, with each element
being the key-value pair arguments to the location.
}
}
\description{
List known locations.
}
\seealso{
\link{orderly_location_pull_metadata}, which can
update your outpack index with metadata from any of the
locations listed here.
}
