% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{orderly_location_pull_packet}
\alias{orderly_location_pull_packet}
\title{Pull one or more packets from a location}
\usage{
orderly_location_pull_packet(
  ...,
  options = NULL,
  recursive = NULL,
  root = NULL
)
}
\arguments{
\item{...}{Arguments passed through to
\link{orderly_search}. In the special case where the first
argument is a character vector of ids \emph{and} there are no named
dot arguments, then we interpret this argument as a vector of
ids directly. Be careful here, your query may pull a lot of data
\itemize{
\item in particular, passing \code{NULL} will match everything that every
remote has!
}}

\item{options}{Options passed to \link{orderly_search}.
The option \code{allow_remote} must be \code{TRUE} as otherwise no packet could
possibly be pulled, so an error is thrown if this is FALSE.}

\item{recursive}{If non-NULL, a logical, indicating if we should
recursively pull all packets that are referenced by the packets
specified in \code{id}.  This might copy a lot of data!  If \code{NULL},
we default to the value given by the the configuration option
\code{require_complete_tree}.}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working
directory. This function does not require that the directory is
configured for orderly, and can be any \code{outpack} root (see
\link{orderly_init} for details).}
}
\value{
Invisibly, the ids of packets that were pulled
}
\description{
Pull one or more packets (including all their files) into this
archive from one or more of your locations. This will make files
available for use as dependencies (e.g., with
\link{orderly_dependency}).
}
\details{
It is possible that it will take a long time to pull packets, if
you are moving a lot of data or if you are operating over a slow
connection.  Cancelling and resuming a pull should be fairly
efficient, as we keep track of files that are copied over even in
the case of an interrupted pull.
}
