% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{orderly_location_push}
\alias{orderly_location_push}
\title{Push tree to location}
\usage{
orderly_location_push(packet_id, location, root = NULL, locate = TRUE)
}
\arguments{
\item{packet_id}{One or more packets to push to the server}

\item{location}{The name of a location to push to (see
\link{orderly_location_list} for possible values).}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function does not require that the
directory is configured for orderly, and can be any \code{outpack}
root (see \link{orderly_init} for details).}

\item{locate}{Logical, indicating if the root should be searched
for.  If \code{TRUE}, then we looks in the directory given for \code{root}
(or the working directory if \code{NULL}) and then up through its
parents until it finds an \code{.outpack} directory or
\code{orderly_config.yml}}
}
\value{
Invisibly, details on the information that was actually
moved (which might be more or less than what was requested,
depending on the dependencies of packets and what was already
known on the other location).
}
\description{
Push tree to location. This function works out what packets are
not known at the location and then what files are required to
create them. It then pushes all the files required to build all
packets and then pushes the missing metadata to the server. If the
process is interrupted it is safe to resume and will only transfer
files and packets that were missed on a previous call.
}
