% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_logging.R
\name{orderly_log_read}
\alias{orderly_log_read}
\title{Read packet logs}
\usage{
orderly_log_read(id, root = NULL)
}
\arguments{
\item{id}{The identifier of the packet to read logs for}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
A \link{data.frame} of log information
}
\description{
Read logs of a previously-run packet. This will fail if the logs
are not available locally (e.g., if a packet was run remotely, or
if you have manually deleted the logs).
}
\details{
Every time that outpack runs a packet, it saves a log at a path
\code{logs.json}; this function reads that file and deserialises it
into a \link{data.frame} for ease of access from R.
}
