% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-orderly2.R
\name{orderly_migrate_source_from_orderly2}
\alias{orderly_migrate_source_from_orderly2}
\title{Migrate orderly2 sources}
\usage{
orderly_migrate_source_from_orderly2(path = ".", dry_run = FALSE)
}
\arguments{
\item{path}{Path to the repo.  We will not change anything here
unless the path is under source control, and unless the git
status is "up to date" (i.e., no local unsaved modifications or
untracked files).  It is recommended to run this against a fresh
clone.}

\item{dry_run}{Logical, indicating if no changes would be made,
but just print information about the changes that would be made.}
}
\value{
Primarily called for side effects, but returns (invisibly)
\code{TRUE} if any changes were made, \code{FALSE} otherwise.
}
\description{
Migrate source code from orderly2 (version 1.99.81 and previous)
to refer to orderly (version 1.99.82 and subsequent).  This is a
one-off and will not exist in the package for long, get it while
it's hot.  This function does a very simple-minded replacement of
\code{orderly2} with \code{orderly} in R files (extension \code{.R} or \code{.r}),
markdown (\code{.md}), R markdown (\code{.Rmd}) and quarto (\code{.qmd}).  It
requires a clean git status before it is run, and will be best to
run against a fresh clone of a repository.  After running, review
changes (if any) with \verb{git diff} and then commit.
}
