% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{orderly_parameters}
\alias{orderly_parameters}
\title{Declare orderly parameters}
\usage{
orderly_parameters(...)
}
\arguments{
\item{...}{Any number of parameters}
}
\value{
Undefined
}
\description{
Declare orderly parameters. You should only have one call to this
within your file, though this is not enforced! Typically you'd put
it very close to the top, though the order does not really matter.
Parameters are scalar atomic values (e.g. a string, number or
boolean) and defaults must be present literally (i.e., they may
not come from a variable itself). Provide \code{NULL} if you do not
have a default, in which case this parameter will be required.
}
