% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{orderly_parse}
\alias{orderly_parse}
\title{Parse the orderly entrypoint script}
\usage{
orderly_parse(path, exprs = NULL, filename = NULL)
}
\arguments{
\item{path}{Path to \verb{orderly_*} script}

\item{exprs}{Parsed AST from \verb{orderly_*} script}

\item{filename}{Name of \verb{orderly_*} file to include in metadata, required if
calling with \code{exprs}. If called with a \code{path} this can be NULL.}
}
\value{
Parsed orderly entrypoint script
}
\description{
For expert use only.
}
\details{
Takes either a path to the orderly entrypoint script or
the parsed AST from an orderly script, parses details
of any calls to the orderly_ in-script functions into intermediate
representation for downstream use. Also validates that any calls to
\verb{orderly_*} in-script functions are well-formed.
}
