% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{orderly_plugin_register}
\alias{orderly_plugin_register}
\title{Register a plugin}
\usage{
orderly_plugin_register(plugin, name)
}
\arguments{
\item{plugin}{The plugin}

\item{name}{The name, typically the package name (later this will
be made optional)}
}
\value{
Nothing, this is called for the side effect of registering
a plugin
}
\description{
Register a plugin
}
