% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_explain.R
\name{orderly_query_explain}
\alias{orderly_query_explain}
\title{Explain a query}
\usage{
orderly_query_explain(
  ...,
  parameters = NULL,
  envir = parent.frame(),
  options = NULL,
  root = NULL
)
}
\arguments{
\item{...}{Arguments passed through to \link{orderly_query},
perhaps just a query expression}

\item{parameters}{Optionally, a named list of parameters to substitute
into the query (using the \verb{this:} prefix)}

\item{envir}{Optionally, an environment to substitute into the
query (using the \verb{environment:} prefix). The default here is to
use the calling environment, but you can explicitly pass this in
if you want to control where this lookup happens.}

\item{options}{Optionally, a \link{orderly_search_options}
object for controlling how the search is performed, and which
packets should be considered in scope. If not provided, default
options are used (i.e., \code{orderly2::orderly_search_options()})}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function does not require that the
directory is configured for orderly, and can be any \code{outpack}
root (see \link{orderly_init} for details).}
}
\value{
An object of class \code{orderly_query_explain}, which can be
inspected (contents subject to change) and which has a print
method which will show a user-friendly summary of the query
result.
}
\description{
Explain how a query has or has not matched. This is experimental
and the output will change. At the moment, it can tell you why a
query matches, or if fails to match based on one of a number of
\code{&&}-ed together clauses.
}
