% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_search.R
\name{orderly_search}
\alias{orderly_search}
\title{Query outpack's database}
\usage{
orderly_search(
  expr,
  name = NULL,
  scope = NULL,
  subquery = NULL,
  parameters = NULL,
  envir = parent.frame(),
  options = NULL,
  root = NULL
)
}
\arguments{
\item{expr}{The query expression. A \code{NULL} expression matches everything.}

\item{name}{Optionally, the name of the packet to scope the query on. This
will be intersected with \code{scope} arg and is a shorthand way of running
\code{scope = list(name = "name")}}

\item{scope}{Optionally, a scope query to limit the packets
searched by \code{pars}}

\item{subquery}{Optionally, named list of subqueries which can be
referenced by name from the \code{expr}.}

\item{parameters}{Optionally, a named list of parameters to substitute
into the query (using the \verb{this:} prefix)}

\item{envir}{Optionally, an environment to substitute into the
query (using the \verb{environment:} prefix). The default here is to
use the calling environment, but you can explicitly pass this in
if you want to control where this lookup happens.}

\item{options}{Optionally, a \link{orderly_search_options}
object for controlling how the search is performed, and which
packets should be considered in scope. If not provided, default
options are used (i.e., \code{orderly2::orderly_search_options()})}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working
directory. This function does not require that the directory is
configured for orderly, and can be any \code{outpack} root (see
\link{orderly_init} for details).}
}
\value{
A character vector of matching ids. In the case of no
match from a query returning a single value (e.g., \code{latest(...)}
or \code{single(...)}) this will be a character missing value
(\code{NA_character_})
}
\description{
Evaluate a query against the outpack database, returning a vector
of matching packet ids.  Note that by default this only searches
through packets that are unpacked and available for direct use on
this computer; to search within packets known to other locations
(and that we might know about via their metadata) you will need to
use the \code{options} argument.
}
