% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{orderly_shared_resource}
\alias{orderly_shared_resource}
\title{Copy shared resources into a packet directory}
\usage{
orderly_shared_resource(...)
}
\arguments{
\item{...}{Named arguments corresponding to shared resources to
copy. The name will be the destination filename, while the value
is the filename within the shared resource directory.}
}
\value{
Invisibly, a character vector of the names of files that
were copied over (the name that they will have in the running
report, not the source name). As for
\link{orderly_resource}, do not rely on the ordering where
directory expansion was performed.
}
\description{
Copy shared resources into a packet directory. You can use this to
share common resources (data or code) between multiple packets.
Additional metadata will be added to keep track of where the files
came from.  Using this function requires the shared resources
directory \verb{shared/} exists at the orderly root; an error will be
raised if this is not configured when we attempt to fetch files.
}
