% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{orderly_zip_export}
\alias{orderly_zip_export}
\title{Export packets as a zip file.}
\usage{
orderly_zip_export(path, packets, root = NULL, locate = TRUE)
}
\arguments{
\item{path}{the path where the zip file will be created}

\item{packets}{One or more packets to export}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function does not require that the
directory is configured for orderly, and can be any \code{outpack}
root (see \link{orderly_init} for details).}

\item{locate}{Logical, indicating if the root should be searched
for.  If \code{TRUE}, then we looks in the directory given for \code{root}
(or the working directory if \code{NULL}) and then up through its
parents until it finds an \code{.outpack} directory or
\code{orderly_config.yml}}
}
\value{
Invisibly, the path to the zip file
}
\description{
The packets can be imported into a different repository using the
\link{orderly_zip_import} function.
}
\details{
This is useful as one-time way to publish your results, for example as an
artefact accompanying a paper. For back-and-forth collaboration, a shared
location should be preferred, as this offers more flexibility.
}
