% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_checksum.R
\name{outpack_checksum}
\alias{outpack_checksum}
\title{Compute packet metadata checksum}
\usage{
outpack_checksum(hash_algorithm = NULL, root = NULL)
}
\arguments{
\item{hash_algorithm}{The hash algorithm to use, defaulting to the
root's hash algorithm.}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
A hash string
}
\description{
Compute a checksum over all known packet metadata. This can be
used to check whether a cache of all metadata needs refreshing (or
agrees between two roots) without having to read a lot of data. It
is unlikely that most users will call this function.
}
\details{
This function matches the behaviour (and hash) of the outpack
server. We may expand the behaviour or either or both of these to
include a similar checksum that includes location data.
}
