% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_helpers.R
\name{outpack_copy_files}
\alias{outpack_copy_files}
\title{Copy files from a packet}
\usage{
outpack_copy_files(
  id,
  files,
  dest,
  allow_remote = FALSE,
  overwrite = TRUE,
  envir = parent.frame(),
  root = NULL
)
}
\arguments{
\item{id}{Optionally, an outpack id via \link{outpack_id}. If
not given a new id will be generated.}

\item{files}{A named character vector of files; the name
corresponds to the name within the current packet, while the
value corresponds to the name within the upstream packet. If you
want to import a directory of files from a packet, you must
refer to the source with a trailing slash (e.g., \code{c(here = "there/")}), which will create the local directory \code{here/...}
with files from the upstream packet directory \verb{there/}. If you
omit the slash then an error will be thrown suggesting that you
add a slash if this is what you intended.}

\item{dest}{The directory to copy into}

\item{allow_remote}{Logical, indicating if we should attempt to
retrieve the file from any remote location if it cannot be found
locally. If the file is large, this may take some time depending
on the speed of the connection. If you use a file store, note
that this does add the downloaded file into your file store,
though associated with no packet so that it is subject to
garbage collection (once we write support for that).}

\item{overwrite}{Overwrite files at the destination; this is
typically what you want, but set to \code{FALSE} if you would prefer
that an error be thrown if the destination file already exists.}

\item{envir}{An environment into which string interpolation
may happen (TODO: docs!). The default here is to use the calling
environment, whcih is typically reasonableb, but may need
changing in programmatic use.}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Nothing, invisibly. Primarily called for its side effect
of copying files from a packet into the directory \code{dest}
}
\description{
Copy files from a packet to anywhere. Similar to
\link{outpack_packet_use_dependency} except that this is not
used in an active packet context. You can use this function to
pull files from an outpack root to a directory outside of the
control of outpack, for example.
}
\details{
There are different ways that this might fail (or recover from
failure):
\itemize{
\item if \code{id} is not known in the metadata store (not known because
it's not unpacked but also not known to be present in some other
remote) then this will fail because it's impossible to resolve
the files. Consider refreshing the metadata with
\link{outpack_location_pull_metadata} to refresh this.
\item if the \code{id} is not unpacked \emph{and} no local copy of the files
referred to can be found, we error by default (but see the next
option). However, sometimes the file you refer to might also be
present because you have downloaded a packet that depended on
it, or because the content of the file is unchanged because from
some other packet version you have locally.
\item if the \code{id} is not unpacked, there is no local copy of the file
and if \code{allow_remote} is \code{TRUE} we will try and request the file
from whatever remote would be selected by
\link{outpack_location_pull_packet} for this packet.
}

Note that empty directories might be created on failure.
}
