% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_root.R
\name{outpack_init}
\alias{outpack_init}
\title{Create outpack root}
\usage{
outpack_init(
  root,
  path_archive = "archive",
  use_file_store = FALSE,
  require_complete_tree = FALSE,
  logging_console = TRUE,
  logging_threshold = "info"
)
}
\arguments{
\item{root}{Path to use.  This path may exist, but it is an error
to call this on a path that has already been initialised.}

\item{path_archive}{Path to the archive directory, used to store
human-readable copies of packets.  If \code{NULL}, no such copy is
made, and \code{file_store} must be \code{TRUE}}

\item{use_file_store}{Logical, indicating if we should use a
content-addressable file-store as the source of truth for
packets.  If \code{archive} is non-\code{NULL}, the file-store will be
used as the source of truth and the duplicated files in archive
exist only for convenience.}

\item{require_complete_tree}{Logical, indicating if we require a
complete tree of packets.  This currently affects
\link{outpack_location_pull_packet}, by requiring that it
always operates in recursive mode.  This is \code{FALSE} by default,
but set to \code{TRUE} if you want your archive to behave well as a
location; if \code{TRUE} you will always have all the packets that
you hold metadata about.}

\item{logging_console}{Logical, indicating if we should log to the
console. If \code{TRUE}, then many operations will produce
informational output; set to \code{FALSE} to prevent this.}

\item{logging_threshold}{The degree of verbosity; one of \code{info},
\code{debug} or \code{trace} in increasing order of verbosity.}
}
\value{
Invisibly, an \code{outpack_root} object; these will change in
future versions!
}
\description{
Create a new outpack root.
}
