% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_location.R
\name{outpack_location_pull_metadata}
\alias{outpack_location_pull_metadata}
\title{Pull metadata from a location}
\usage{
outpack_location_pull_metadata(location = NULL, root = NULL)
}
\arguments{
\item{location}{The name of a location to pull from (see
\link{outpack_location_list} for possible values).  If not
given, pulls from all locations.  The "local" and "orphan"
locations are always up to date and pulling metadata from them
does nothing.}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Nothing
}
\description{
Pull metadata from a location, updating the index.  This should
always be relatively quick as it updates only small files that
contain information about what can be found in remote packets.
}
