% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_location.R
\name{outpack_location_push}
\alias{outpack_location_push}
\title{Push tree to location}
\usage{
outpack_location_push(packet_id, location, root = NULL)
}
\arguments{
\item{packet_id}{One or more packets to push to the server}

\item{location}{The name of a location to push to (see
\link{outpack_location_list} for possible values).}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Invisibly, details on the information that was actually
moved (which might be more or less than what was requested,
depending on the dependencies of packets and what was already
known on the other location).
}
\description{
Push tree to location. This function works out what packets are
not known at the location and then what files are required to
create them. It then pushes all the files required to build all
packets and then pushes the missing metadata to the server. If the
process is interrupted it is safe to resume and will only transfer
files and packets that were missed on a previous call.
}
