% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_location.R
\name{outpack_location_remove}
\alias{outpack_location_remove}
\title{Remove a location}
\usage{
outpack_location_remove(name, root = NULL)
}
\arguments{
\item{name}{The short name of the location.
Cannot remove \code{local} or \code{orphan}}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Nothing
}
\description{
Remove an existing location. Any packets from this location
will now be associated with the 'orphan' location instead.
}
