% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_location.R
\name{outpack_location_rename}
\alias{outpack_location_rename}
\title{Rename a location}
\usage{
outpack_location_rename(old, new, root = NULL)
}
\arguments{
\item{old}{The current short name of the location.
Cannot rename \code{local} or \code{orphan}}

\item{new}{The desired short name of the location.
Cannot be one of \code{local} or \code{orphan}}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Nothing
}
\description{
Rename an existing location
}
