% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_metadata.R
\name{outpack_metadata_read}
\alias{outpack_metadata_read}
\title{Read outpack metadata json file}
\usage{
outpack_metadata_read(path)
}
\arguments{
\item{path}{Path to the json file}
}
\value{
A list of outpack metadata; see the schema for details. In
contrast to reading the json file directly with
\code{jsonlite::fromJSON}, this function will take care to convert
scalar and length-one vectors into the expected types.
}
\description{
Low-level function for reading metadata and deserialising it. This
function can be used to directly read a metadata json file without
reference to a root which contains it. It may be useful in the
context of reading a metadata file written out as part of a failed
run.
}
