% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_packet.R
\name{outpack_packet_file_mark}
\alias{outpack_packet_file_mark}
\alias{outpack_packet_file_list}
\title{Mark files during packet run}
\usage{
outpack_packet_file_mark(packet, files, status)

outpack_packet_file_list(packet)
}
\arguments{
\item{packet}{Optionally, an explicitly-passed packet; see
\code{\link[=outpack_packet_start]{outpack_packet_start()}} for details.}

\item{files}{A character vector of relative paths}

\item{status}{A status to mark the file with. Must be "immutable"
or "ignored"}
}
\value{
Depending on function
\itemize{
\item \code{outpack_packet_file_mark} returns nothing
\item \code{outpack_packet_file_list} returns a \link{data.frame} with columns
\code{path} and \code{status} (\code{immutable}, \code{ignored} or \code{unknown})
}
}
\description{
Mark file within an in-progress packet. This will store the hash
of the file within the internal outpack structures and force an
error if the file is changed or deleted later.  The function
\code{\link[=outpack_packet_file_list]{outpack_packet_file_list()}} will report on which files
are marked (or unmarked) within the directory.
}
