% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_deparse.R
\name{outpack_query_format}
\alias{outpack_query_format}
\title{Format outpack query for displaying to users. It will typically be
easier to use the \link{format} method of \code{outpack_query} objects.}
\usage{
outpack_query_format(query, subquery = NULL)
}
\arguments{
\item{query}{The outpack query to print}

\item{subquery}{Optionally a named list of subqueries - if given,
each must be a valid deparseable subquery (i.e., a literal or
language object).  Note that you must not provide an
already-deparsed query here or it will get quoted!}
}
\value{
Query expression as a string
}
\description{
Format outpack query for displaying to users. It will typically be
easier to use the \link{format} method of \code{outpack_query} objects.
}
\examples{
outpack::outpack_query_format(quote(name == "example"))
outpack::outpack_query_format(
  quote(usedby({A})),
  subquery = list(A = quote(latest(name == "a"))))

format(outpack::outpack_query("latest", name = "a"))
}
