% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_root.R
\name{outpack_root_open}
\alias{outpack_root_open}
\title{Open outpack root}
\usage{
outpack_root_open(path, locate = TRUE)
}
\arguments{
\item{path}{The path to look for the root; must be an existing
directory. Use \code{NULL} (or \code{.} or \code{getwd()}) for the current
directory.}

\item{locate}{Logical, indicating if we should look in parent
directories until the root is found (similar behaviour to how
\code{git} can find its root directory from any directory below the
root).}
}
\value{
An \code{outpack_root} object; treat this as an opaque object
for now.
}
\description{
Open an existing outpack root. This returns a "root" object, which
can be passed through to various outpack functions. The root
object is the same as that returned by
\link{outpack_init} and will be documented once the
interface stabilises.
}
