% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_search.R
\name{outpack_search_options}
\alias{outpack_search_options}
\title{Packet search options}
\usage{
outpack_search_options(
  location = NULL,
  allow_remote = FALSE,
  pull_metadata = FALSE
)
}
\arguments{
\item{location}{Optional vector of locations to pull from. We
might in future expand this to allow wildcards, exceptions, or
numeric values corresponding to the location priority (and then
it's possible we'll change the name).}

\item{allow_remote}{Logical, indicating if we should allow
packets to be found that are not currently unpacked (i.e., are
known only to a location that we have metadata from). If this is
\code{TRUE}, then inconjunction with
\link{outpack_packet_use_dependency} you might pull a large
quantity of data.}

\item{pull_metadata}{Logical, indicating if we should pull
metadata immediately before the search. If \code{location} is
given, then we will pass this through to
\link{outpack_location_pull_metadata} to filter locations to
update.  If pulling many packets in sequence, you \emph{will} want to
update this option to \code{FALSE} after the first pull.}
}
\value{
An object of class \code{outpack_search_options} which should
not be modified after creation (but see note about \code{pull_metadata})
}
\description{
Options for controlling how packet searches are carried out, for
example via \link{outpack_search} and
\link{outpack_packet_use_dependency}. The details here are
never included in the metadata alongside the query (that is,
they're not part of the query even though they affect it).
}
