% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{orderly_config_set}
\alias{orderly_location_pull}
\alias{orderly_location_remove}
\alias{orderly_cleanup}
\alias{orderly_plugin_register}
\alias{orderly_config}
\alias{orderly_copy_files}
\alias{orderly_run_info}
\alias{orderly_example_show}
\alias{orderly_init}
\alias{orderly_location_fetch_metadata}
\alias{orderly_new}
\alias{orderly_dependency}
\alias{orderly_gitignore_update}
\alias{orderly_query_explain}
\alias{orderly_plugin_add_metadata}
\alias{orderly_run}
\alias{orderly_prune_orphans}
\alias{orderly_resource}
\alias{orderly_location_rename}
\alias{orderly_parse_expr}
\alias{orderly_parameters}
\alias{orderly_location_add_packit}
\alias{orderly_location_add_http}
\alias{orderly_strict_mode}
\alias{orderly_location_pull_metadata}
\alias{orderly_metadata}
\alias{orderly_comparison_explain}
\alias{orderly_cleanup_status}
\alias{orderly_plugin_context}
\alias{orderly_metadata_read}
\alias{orderly_validate_archive}
\alias{orderly_example}
\alias{orderly_query}
\alias{orderly_search}
\alias{orderly_hash_data}
\alias{orderly_hash_file}
\alias{orderly_location_add}
\alias{orderly_description}
\alias{orderly_location_push}
\alias{orderly_artefact}
\alias{orderly_metadata_extract}
\alias{orderly_location_add_path}
\alias{orderly_location_pull_packet}
\alias{orderly_location_list}
\alias{orderly_parse_file}
\alias{orderly_shared_resource}
\alias{orderly_compare_packets}
\alias{orderly_list_src}
\alias{orderly_interactive_set_search_options}
\alias{orderly_search_options}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{orderly}{\code{\link[orderly]{orderly_artefact}}, \code{\link[orderly]{orderly_cleanup}}, \code{\link[orderly:orderly_cleanup]{orderly_cleanup_status}}, \code{\link[orderly]{orderly_compare_packets}}, \code{\link[orderly]{orderly_comparison_explain}}, \code{\link[orderly]{orderly_config}}, \code{\link[orderly]{orderly_config_set}}, \code{\link[orderly]{orderly_copy_files}}, \code{\link[orderly]{orderly_dependency}}, \code{\link[orderly]{orderly_description}}, \code{\link[orderly]{orderly_example}}, \code{\link[orderly]{orderly_example_show}}, \code{\link[orderly]{orderly_gitignore_update}}, \code{\link[orderly:orderly_hash]{orderly_hash_data}}, \code{\link[orderly:orderly_hash]{orderly_hash_file}}, \code{\link[orderly]{orderly_init}}, \code{\link[orderly]{orderly_interactive_set_search_options}}, \code{\link[orderly]{orderly_list_src}}, \code{\link[orderly]{orderly_location_add}}, \code{\link[orderly:orderly_location_add]{orderly_location_add_http}}, \code{\link[orderly:orderly_location_add]{orderly_location_add_packit}}, \code{\link[orderly:orderly_location_add]{orderly_location_add_path}}, \code{\link[orderly]{orderly_location_fetch_metadata}}, \code{\link[orderly]{orderly_location_list}}, \code{\link[orderly]{orderly_location_pull}}, \code{\link[orderly:orderly-deprecated]{orderly_location_pull_metadata}}, \code{\link[orderly:orderly-deprecated]{orderly_location_pull_packet}}, \code{\link[orderly]{orderly_location_push}}, \code{\link[orderly]{orderly_location_remove}}, \code{\link[orderly]{orderly_location_rename}}, \code{\link[orderly]{orderly_metadata}}, \code{\link[orderly]{orderly_metadata_extract}}, \code{\link[orderly]{orderly_metadata_read}}, \code{\link[orderly]{orderly_new}}, \code{\link[orderly]{orderly_parameters}}, \code{\link[orderly:orderly_parse_file]{orderly_parse_expr}}, \code{\link[orderly]{orderly_parse_file}}, \code{\link[orderly]{orderly_plugin_add_metadata}}, \code{\link[orderly]{orderly_plugin_context}}, \code{\link[orderly]{orderly_plugin_register}}, \code{\link[orderly]{orderly_prune_orphans}}, \code{\link[orderly]{orderly_query}}, \code{\link[orderly]{orderly_query_explain}}, \code{\link[orderly]{orderly_resource}}, \code{\link[orderly]{orderly_run}}, \code{\link[orderly]{orderly_run_info}}, \code{\link[orderly]{orderly_search}}, \code{\link[orderly]{orderly_search_options}}, \code{\link[orderly]{orderly_shared_resource}}, \code{\link[orderly]{orderly_strict_mode}}, \code{\link[orderly]{orderly_validate_archive}}}
}}

