% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore.R
\name{orderly_gitignore_update}
\alias{orderly_gitignore_update}
\title{Update a gitignore file}
\usage{
orderly_gitignore_update(
  name,
  prompt = "if_manually_created",
  root = NULL,
  locate = TRUE
)
}
\arguments{
\item{name}{The name of the gitignore file to update, or the
string "(root)"}

\item{prompt}{Conditions under which you will be prompted before
writing anything to a file (in a non-interactive session, we
error instead). Options here are
\itemize{
\item \code{if_manually_created}": (the default), which prompts if a
.gitignore file exists with content that does not include
special orderly2 markers
\item \code{if_new_file}: if we would create a new file
\item \code{always}: on any change to contents
\item \code{never}: never prompt, just write the changes
}}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function \strong{does} require that the
directory is configured for orderly, and not just outpack (see
\link{orderly_init} for details).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{.outpack} directory}
}
\value{
Nothing, called for its side effects
}
\description{
Update a gitignore, which is useful to prevent accidentally
comitting files to source control that are generated. This
includes artefacts, shared resources and dependencies (within a
report directory) or at the global level all the contents of the
\code{.outpack} directory, the draft folder and the archive directory.
}
\details{
If this function fails with a message \verb{Can't edit '.gitignore', markers are corrupted}, then look for the special markers within
the \code{.gitignore} file.  It should look like

\if{html}{\out{<div class="sourceCode">}}\preformatted{# ---VVV--- added by orderly ---VVV----------------
# Don't manually edit content between these markers
... patterns
# ---^^^--- added by orderly ---^^^----------------
}\if{html}{\out{</div>}}

We can't edit the file if:
\itemize{
\item any of these lines appears more than once in the file
\item there is anything between the first two lines
\item they are not in this order
}

If you get the error message, search and remove these lines and
rerun.
}
