% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{orderly_init}
\alias{orderly_init}
\title{Initialise an orderly repository}
\usage{
orderly_init(path, ...)
}
\arguments{
\item{path}{The path to initialise the repository at.  If the
repository is already initialised, this operation does nothing.}

\item{...}{Additional arguments passed through to
\link{outpack_init}, if (and only if) the outpack
repository is not already created.}
}
\value{
An orderly root object, invisibly. Typically this is
called only for its side effect.
}
\description{
Initialise an empty orderly repository. An orderly repository is
also an outpack repository, so most of the work here is done by
\code{\link[=outpack_init]{outpack_init()}}, however, you should call this
function!  You can also turn any outpack repository into an
orderly one, adding appropriate bits of configuration into it,
though this function.
}
