% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_location.R
\name{orderly_location_remove}
\alias{orderly_location_remove}
\title{Remove a location}
\usage{
orderly_location_remove(name, root = NULL, locate = TRUE)
}
\arguments{
\item{name}{The short name of the location.
Cannot remove \code{local} or \code{orphan}}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function does not require that the
directory is configured for orderly, and can be any \code{outpack}
root (see \link{orderly_init} for details).}

\item{locate}{Logical, indicating if the root should be searched
for.  If \code{TRUE}, then we looks in the directory given for \code{root}
(or the working directory if \code{NULL}) and then up through its
parents until it finds an \code{.outpack} directory or
\code{orderly_config.yml}}
}
\value{
Nothing
}
\description{
Remove an existing location. Any packets from this location
will now be associated with the 'orphan' location instead.
}
