% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpack_logging.R
\name{orderly_log_read}
\alias{orderly_log_read}
\title{Read packet logs}
\usage{
orderly_log_read(id, root = NULL, locate = TRUE)
}
\arguments{
\item{id}{The identifier of the packet to read logs for}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function does not require that the
directory is configured for orderly, and can be any \code{outpack}
root (see \link{orderly_init} for details).}

\item{locate}{Logical, indicating if the root should be searched
for.  If \code{TRUE}, then we looks in the directory given for \code{root}
(or the working directory if \code{NULL}) and then up through its
parents until it finds an \code{.outpack} directory or
\code{orderly_config.yml}}
}
\value{
A \link{data.frame} of log information
}
\description{
Read logs of a previously-run packet. This will fail if the logs
are not available locally (e.g., if a packet was run remotely, or
if you have manually deleted the logs).
}
\details{
Every time that outpack runs a packet, it saves a log at a path
\code{logs.json}; this function reads that file and deserialises it
into a \link{data.frame} for ease of access from R.
}
