% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{orderly_migrate_source_from_orderly2}
\alias{orderly_migrate_source_from_orderly2}
\title{Migrate orderly2 sources}
\usage{
orderly_migrate_source_from_orderly2(path)
}
\arguments{
\item{path}{Path to the repo.  We will not change anything here
unless the path is under source control, and unless the git
status is "up to date" (i.e., no local unsaved modifications or
untracked files).  It is recommended to run this against a fresh
clone.}
}
\value{
Nothing, called for side effects only
}
\description{
Migrate source code from orderly2 (version 1.99.78 and previous)
to refer to orderly (version 1.99.79 and subsequent).  This is a
one-off and will not exist in the package for long, get it while
it's hot.
}
