% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{orderly_new}
\alias{orderly_new}
\title{Create a new report}
\usage{
orderly_new(name, template = NULL, force = FALSE, root = NULL, locate = TRUE)
}
\arguments{
\item{name}{The name of the report}

\item{template}{The template to use. The only acceptable values
for now are \code{NULL} (uses the built-in default) and \code{FALSE} which
suppresses any default content.  We may support customisable
templates in future - let us know if this would be useful.}

\item{force}{Create an orderly file - \verb{<name>.R} within an existing
directory \verb{src/<name>}; this may be useful if you have already
created the directory and some files first but want help
creating the orderly file.}

\item{root}{The path to the root directory, or \code{NULL} (the
default) to search for one from the current working directory if
\code{locate} is \code{TRUE}. This function \strong{does} require that the
directory is configured for orderly, and not just outpack (see
\link{orderly_init} for details).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{.outpack} directory}
}
\value{
Nothing, called for its side effects only
}
\description{
Create a new empty report.
}
