% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{orderly_parse}
\alias{orderly_parse}
\title{Parse the orderly entrypoint script}
\usage{
orderly_parse(entrypoint_script, entrypoint_filename)
}
\arguments{
\item{entrypoint_script}{Path to script or parsed AST from orderly script}

\item{entrypoint_filename}{Name of entrypoint file to include in metadata}
}
\value{
Parsed orderly entrypoint script
}
\description{
For expert use only.
}
\details{
Takes either a path to the orderly entrypoint script or
the parsed AST from an orderly script, parses details
of any calls to the orderly_ in-script functions into intermediate
representation for downstream use. Also validates that any calls to
orderly_ in-script functions are well-formed.
}
