% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{orderly_run}
\alias{orderly_run}
\title{Run a report}
\usage{
orderly_run(
  name,
  parameters = NULL,
  envir = NULL,
  logging_console = NULL,
  logging_threshold = NULL,
  root = NULL,
  locate = TRUE
)
}
\arguments{
\item{name}{Name of the report to run}

\item{parameters}{Parameters passed to the report. A named list of
parameters declared in the \code{orderly.yml}.  Each parameter
must be a scalar character, numeric, integer or logical.}

\item{envir}{The environment that will be used to evaluate the
report script; by default we use the global environment, which
may not always be what is wanted.}

\item{logging_console}{Optional logical, passed through to
\link[outpack:outpack_packet]{outpack::outpack_packet_start} to control printing logs to the
console, overriding any default configuration set at the root.}

\item{logging_threshold}{Optional logging threshold, passed through to
\link[outpack:outpack_packet]{outpack::outpack_packet_start} to control the amount of detail
in logs printed during running, overriding any default
configuration set at the root. If given, must be one of \code{info},
\code{debug} or \code{trace} (in increasing order of
verbosity). Practically this has no effect at present as we've
not added any fine-grained logging.}

\item{root}{The path to an orderly root directory, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}.}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{.outpack} directory}
}
\value{
The id of the created report (a string)
}
\description{
Run a report.  This will create a new directory in
\verb{drafts/<reportname>}, copy your declared resources there, run
your script and check that all expected artefacts were created.
}
